requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
boto3==1.42.96
botocore==1.42.96
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==47.0.0
Django==5.2.13
django-allauth==65.16.1
django-cors-headers==4.9.0
django-environ==0.13.0
django-extensions==4.1
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.17.1
drf-spectacular==0.29.0
fido2==2.2.0
gunicorn==25.3.0
idna==3.13
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==26.2
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
s3transfer==0.16.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wheel==0.47.0
whitenoise==6.12.0
